Move schemas into schema file
